requirements.txt
python-telegram-bot>=22
dateparser
httpx>=0.24
python-dotenv